🏷️ Remove duplicate type definitions for Nunjucks API
