+++ types.node_modules/nunjucks/src/lib.d.ts
@@ -5,8 +5,12 @@
 /**
  * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L47-L136 Source}
  */
-export class TemplateError extends Error {
-    constructor(message: string | Error, lineno?: number, colno?: number);
+export class TemplateError<
+    TMessage extends string | Error = string | Error,
+    TLineno extends number | undefined = number | undefined,
+    TColno extends number | undefined = number | undefined
+> extends Error {
+    constructor(message: TMessage, lineno?: TLineno, colno?: TColno);
 
     /**
      * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L68-L70 Source}
@@ -16,20 +20,21 @@ export class TemplateError extends Error {
     /**
      * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L93-L95 Source}
      */
-    readonly cause: Error | undefined;
+    readonly cause: TMessage extends Error ? TMessage : undefined;
 
     /**
      * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L97 Source}
      */
-    lineno: number | undefined;
+    lineno: TLineno;
 
     /**
      * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L98 Source}
      */
-    colno: number | undefined;
+    colno: TColno;
 
     /**
      * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L99 Source}
+     * {@link https://github.com/mozilla/nunjucks/blob/v3.2.1/nunjucks/src/lib.js#L120 Source}
      */
     firstUpdate: boolean;
